+++ integrations/pvscale/configure.php
@@ -5,10 +5,7 @@ require_once(__ROOT__.'/inc/sec.php');
 require_once(__ROOT__.'/inc/opendb.php');
 require_once(__ROOT__.'/inc/settings.php');
 
-if ($role !== 1){
-  echo json_encode(['success' => false, 'error' => 'Not authorised']);
-  return;
-}
+
 ?>
 <div class="card-body">
   <div id="scmsg"></div>