+++ pages/settings.php
@@ -21,9 +21,11 @@ require_once(__ROOT__.'/func/php-settings.php');
         <li><a href="#templates" id="templates_tab" role="tab" data-bs-toggle="tab">HTML Templates</a></li>
         <li><a href="#sds" id="sds_tab" role="tab" data-bs-toggle="tab">SDS Settings</a></li>
         <li><a href="#brand" id="brand_tab" role="tab" data-bs-toggle="tab">Branding</span></a></li>
+        <?php if($system_settings['INTEGRATIONS_enable'] === '1') {?>
+          <li><a href="#integrations" id="integrations_tab" role="tab" data-bs-toggle="tab">Integrations</a></li>
+        <?php } ?>
         <?php if($role === 1){?>
           <li><a href="#maintenance" id="maintenance_tab" role="tab" data-bs-toggle="tab">Maintenance</a></li>
-          <li><a href="#integrations" id="integrations_tab" role="tab" data-bs-toggle="tab">Integrations</a></li>
           <li><a href="#syslogs" id="logs_tab" role="tab" data-bs-toggle="tab">System logs</a></li>
         <?php } ?>
         <li><a href="#api" id="api_tab" role="tab" data-bs-toggle="tab">API</a></li>
@@ -121,14 +123,16 @@ require_once(__ROOT__.'/func/php-settings.php');
           <div class="loader-text"></div>
         </div>
       </div>
-      <?php if($role === 1){?>  
-      <div id="maintenance">
+      <?php if($system_settings['INTEGRATIONS_enable'] === '1') {?>
+      <div id="integrations">
         <div class="loader-center">
           <div class="loader"></div>
           <div class="loader-text"></div>
         </div>
       </div>
-      <div id="integrations">
+      <?php } ?>
+      <?php if($role === 1){?>  
+      <div id="maintenance">
         <div class="loader-center">
           <div class="loader"></div>
           <div class="loader-text"></div>